==> code/01_fare_cleaning_EDA.ipynb <==
## Knowledge-based Recommender System

import pandas as pd
import numpy as np
from scipy import sparse
from sklearn.metrics.pairwise import pairwise_distances

import matplotlib as plt

import re

fl = pd.read_csv('../data/consumer_airfare_report_US.csv')

fl.head()

fl.shape

fl.columns

fl['Year'].unique()

fl.drop(columns= ['tbl', 'Geocoded_City1', 'Geocoded_City2',
       'tbl1apk', 'carrier_lg', 'large_ms', 'fare_lg',
       'carrier_low', 'airportid_1', 'airportid_2', 'airport_1', 'airport_2', 'nsmiles', 'lf_ms'], inplace = True)

fl.head()

fl.shape

fl.isnull().sum()

def camel_to_snake(col):
    return re.sub(r'(?<!^)(?=[A-Z])', '_', col).lower()

fl.columns = [camel_to_snake(col) for col in fl.columns]

citymarketid_1 = City market ID is an identification number assigned by US DOT to identify a city market. Use this field to consolidate airports serving the same city market
citymarketid_2 = 
nsmiles = Non-Stop market miles
passenger = passenger per day
fare = overall average fare
fare_low = Average fare for the lowest carrier

#dropping nulls in fare_low
fl.dropna()

fl['year'].value_counts()

fl.drop(columns= ['quarter', 'citymarketid_1', 'citymarketid_2', 'passengers', 'fare_low'], inplace = True)

#only interested in years 2020-2023 - other fares are more than likely outdated
years_to_keep = [2020, 2021, 2022, 2023]
fl_filtered = fl[fl['year'].isin(years_to_keep)]

fl_filtered.shape

fl_filtered['year'].value_counts()

fl_filtered.head()

fl_filtered['city1'].value_counts

#separate city from state
fl_filtered[['city1', 'state1']] = fl.loc[:, 'city1'].str.split(', ',  n=1, expand=True)
fl_filtered[['city2', 'state2']] = fl.loc[:, 'city2'].str.split(', ', n=1, expand=True)

fl_filtered.isnull().sum()

fl_filtered.shape

fl_filtered.to_csv('../data/flight_data_clean.csv', index=False) 

fl_filtered['city1'].unique()

fl_filtered['fare'].value_counts()

#keeping only city2 that matches cities of airbnb
cities_to_keep = ['Austin', 'Boston', 'Chicago', 'Columbus', 'Denver', 'Ft Lauderdale', 
    'honolulu', 'Jersey City', 'Las Vegas', 'Los Angeles', 'Nashville', 
    'New Orleans', 'New York City', 'Portland', 'Rhode Island', 
    'San Diego', 'San Francisco', 'Seattle', 'Minneapolis', 'Washington']              

filtered_df = fl_filtered[fl_filtered['city2'].isin(cities_to_keep)]

filtered_df.shape

filtered_df.to_csv('../data/flight_data_clean.csv', index=False)

## Pivot Table created but ultimately not used.  

fare_mx = pd.pivot_table(
    fl_filtered,
    values = 'fare',
    index = 'city2',
    columns = 'city1'
)

fare_mx.head()

fare_mx.shape

fl_filtered.columns

Sparse Matrix

fl_filtered['fare'].value_counts()

fare_sparse = sparse.csr_matrix(fare_mx.fillna(0))

print(fare_sparse)

import sys

sys.getsizeof(fare_mx)

sys.getsizeof(fare_sparse)

Cosine Similarity

# distance of 1 is a similarity of 0.
dists = pairwise_distances(fare_sparse, metric='cosine')

print(dists.shape)
dists[:10, :10]

Distances DataFrame

fl_rec = pd.DataFrame(dists, columns=fare_mx.index, index = fare_mx.index)

fl_rec.iloc[:10, :10]

starting_city = fare_mx.index

starting_city[starting_city.str.lower().str.contains('chicago')]

fl_rec['Chicago'].sort_values().iloc[1:].head(10)

starting_city[starting_city.str.lower().str.contains('las vegas')]

fl_rec['Las Vegas'].sort_values().iloc[1:5]
