==> code/02_airbnb_cleaning_EDA.ipynb <==
import pandas as pd
import numpy as np

import re

bnb = pd.read_csv('../data/air-bnb-listings-to-use.csv')

pd.set_option('display.max_rows', 1000)

bnb.head()

bnb.shape

bnb['City'].value_counts()

#changing Broward County to Ft Lauderdale
bnb['City'] = bnb['City'].replace('Broward-county', 'Ft Lauderdale')

#changing Clark County to Las Vegas
bnb['City'] = bnb['City'].replace('Clark-county-nv', 'Las Vegas')

#changing Twin-cities-msa to Twin Cities
bnb['City'] = bnb['City'].replace('Twin-cities-msa', 'Twin Cities')

# mapping to create state column
city_to_state = {
    'Denver': 'CO',
    'San-francisco': 'CA',
    'New-york-city': 'NY',
    'Los Angeles': 'CA',
    'Chicago': 'IL',
    'Miami': 'FL',
    'Austin': 'TX',
    'Seattle': 'WA',
    'Boston': 'MA',
    'Washington-dc': 'DC',
    'Philadelphia': 'PA',
    'Atlanta': 'GA',
    'Dallas': 'TX',
    'Houston': 'TX',
    'Phoenix': 'AZ',
    'Las Vegas': 'NV',
    'Portland': 'OR',
    'San-diego': 'CA',
    'Orlando': 'FL',
    'Columbus': 'OH',
    'Hawaii': 'HI',
    'Oakland': 'CA',
    'Rhode Island': 'RI',
    'Nashville': 'TN',
    'Ft Lauderdale': 'FL',
    'San-mateo-county': 'CA',
    'Jersey-city': 'NJ',
    'Asheville': 'NC',
    'Santa-cruz-county': 'CA',
    'Salem-or': 'OR',
    'New-orleans': 'LA',
    'Cambridge': 'MA',
    'Rhode Island': 'RI',
    'Twin Cities': 'MN'
}

bnb['State'] = bnb['City'].map(city_to_state)

bnb.head()

bnb['State'].value_counts()

bnb['City'].value_counts()

def camel_to_snake(col):
    return re.sub(' ', '_', col).lower()

bnb.columns = [camel_to_snake(col) for col in bnb.columns]

bnb.head()

bnb.columns

def camel_to_snake(city):
    return re.sub('-', ' ', city).title()

bnb['city'] = [camel_to_snake(city) for city in bnb['city']]

bnb['city'].value_counts()

bnb.columns

#drop unnecessary columns
bnb.drop(columns=['name', 'host_id', 'number_of_reviews', 'date_last_review',
       'number_of_reviews_per_month', 'rooms_rent_by_the_host', 'availibility', 'country', 'coordinates', 'location'], inplace=True)

bnb.head()

bnb.isnull().sum()

bnb.shape

# dropping null State
bnb.dropna(subset=['state'], inplace=True)

bnb.shape

bnb.to_csv('../data/air_bnb_data_clean.csv', index=False)
