==> code/05_airfare_cv_cs_notebook.ipynb <==
# Airfare Countvectorizer and Cosine Similarities Performed.

import pandas as pd
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.metrics.pairwise import cosine_similarity

df = pd.read_csv('../data/flight_data_clean.csv')

df.head()

df.shape

df.info()

# Airfare Recommender System.

Followed link provided from Hank to CountVectorize and get Cosine Similarities for a recommender system.  (365datascience.com).  This code provides the similarities between city1 (departure city) and city2 (arrival city) which can be later used with budget input.

#removing duplicates
df.duplicated(subset='city1').sum()

df = df.drop_duplicates(subset='city1')

#checking that duplicates were dropped
df.duplicated(subset='city1').sum()

#dropping columns not needed
df.drop(['year', 'state1', 'state2'], axis=1, inplace=True)

def clean_text(city1):
    result = str(city1).lower()
    return(result.replace(' ', ''))

df['city1'] = df['city1'].apply(clean_text)

df.head()

df['city2'] = df['city2'].str.lower()

df2 = df.drop(['fare'], axis=1)
df2['data'] = df2[df2.columns[0:]].apply( 
    lambda x: ' '.join(x.dropna().astype(str)), axis=1)

print(df2['data'].head())

vectorizer = CountVectorizer()
vectorized = vectorizer.fit_transform(df2['data'])

similarities = cosine_similarity(vectorized)

print(similarities)

df = pd.DataFrame(similarities, columns=df['city1'], index=df['city1']).reset_index()

df.tail()

df.dtypes

input_city = 'richmond'
recommendations = pd.DataFrame(df.nlargest(5,input_city)['city1'])

print(recommendations)

df['city1'].unique()
